+++ CHANGELOG.md
@@ -3,7 +3,6 @@ Changelog
 
 8.28.0
 ------
-* changed: Handling of camera and initial delay moved out of ChoiceWorld's state machine definition
 * added: `_iblrig_tasks_spontaneousBpod` - a spontaneous task that includes Bpod spacers
 * added: save acquisistion log for camera sessions
 * added: save ambient data to Parquet file
@@ -11,6 +10,7 @@ Changelog
 * added: check for remote data folder during hardware validation
 * added: indicate remote rigs in `_iblrig_taskSettings.raw.json`
 * added: `ibl-photometry` dependency
+* changed: Handling of camera and initial delay moved out of ChoiceWorld's state machine definition
 * changed: reduce traceback size in UDP message
 * changed: hardware validation - more solid way to estimate TTL rates
 * changed: clean-up after failed session (terminate Bonsai processes and stop state machine)